pyproject.toml
[project]
name = "svg2ico"
dynamic = ["version"]
description = "Converter from SVG file to Windows .ico file"
authors = [
    {name = "yamakox",email = "yamakox@gmail.com"}
]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD-3-Clause",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta", 
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "cairosvg (>=2.8.0,<3.0.0)",
    "wxpython (>=4.2.3,<5.0.0)",
    "click (>=8.2.0,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
version = "0.0.0"

[tool.poetry.scripts]
svg2ico = "svg2ico:main"

[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry-dynamic-versioning]
enable = true
